Add S3 support per dataset
Currently benchmarking works for individual datasets belonging to the context, i.e., the user has to input datasets when running launch_jobs.py.
Similarly, this is the status for folds, the user has to input folds explicitly.
The next commit will include support for running all folds and all datasets in a context at once.